split field into own file
